requirements.txt
Pillow>=10.0.0
tkinter-tooltip>=2.0.0
numpy>=1.24.0